app/requirements/requirements.txt
arabic-reshaper==2.1.1
asgiref==3.2.10
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cssselect2==0.4.1
defusedxml==0.6.0
Django==3.1.2
django-crum==0.7.5
django-widget-tweaks==1.4.5
et-xmlfile==1.0.1
future==0.18.2
html5lib==1.1
idna==2.10
jdcal==1.4.1
openpyxl==3.0.6
pdfkit==0.6.1
Pillow==7.0.0
pycparser==2.20
Pyphen==0.10.0
python-bidi==0.4.2
pytz==2019.3
requests==2.25.1
six==1.15.0
sqlparse==0.3.1
tinycss2==1.1.0
urllib3==1.26.2
WeasyPrint==52.2
webencodings==0.5.1
XlsxWriter==1.3.7
